Document the new classes and datatypes, change some class names
